Move benchmark and add fake path
